Separated out DreamConn from DreamPicoPort
