+++ CMakeLists.txt
@@ -227,7 +227,7 @@ else()
 	add_executable(${PROJECT_NAME} core/emulator.cpp)
 endif()
 
-set_target_properties(${PROJECT_NAME} PROPERTIES 
+set_target_properties(${PROJECT_NAME} PROPERTIES
 		CMAKE_CXX_STANDARD 20
 		CMAKE_CXX_STANDARD_REQUIRED ON)
 
@@ -481,6 +481,10 @@ if(NOT LIBRETRO)
 			core/sdl/sdl_gamepad.h
 			core/sdl/sdl_keyboard.h
 			core/sdl/sdl_keyboard_mac.h
+			core/sdl/dreamlink.cpp
+			core/sdl/dreamlink.h
+			core/sdl/dreampicoport.cpp
+			core/sdl/dreampicoport.h
 			core/sdl/dreamconn.cpp
 			core/sdl/dreamconn.h)
 
@@ -523,7 +527,7 @@ if(NOT WITH_SYSTEM_ZLIB)
 	# help libzip find the package
 	set(ZLIB_FOUND TRUE)
 	set(ZLIB_INCLUDE_DIR "${CMAKE_CURRENT_SOURCE_DIR}/${ZLIB_RELATIVE_PATH}" "${CMAKE_CURRENT_BINARY_DIR}/${ZLIB_RELATIVE_PATH}")
-	
+
 	cmake_policy(SET CMP0026 OLD)
 	get_target_property(ZLIB_LIBRARY_RELEASE zlibstatic LOCATION)
 	get_target_property(ZLIB_LIBRARY_DEBUG zlibstatic LOCATION_Debug)
